Add 'pkgbuild' verb to fetch PKGBUILDs
Refactor/generalize the download machinery to handle any sort of
urlpath and reuse as much as possible to implement PKGBUILD fetching.
Perhaps this points to a restructuring to either build out the JSON
parsing on top of this new raw request/responce logic, or just move the
it to the frontend (as we do with tarball extraction).

When multiple PKGBUILDs are requested, inject shell comments to
delineate PKGBUILDs.