Extend AURACLE_DEBUG to allow only dumping outbound requests
This makes our e2e testing much more sane, as we can control what and
where our debug output goes. In this case, a prefix of "requests:" in
the AURACLE_DEBUG environment var indicates that we should write
outbound requests to the path which follows the prefix. Use this in our
e2e tests to check what requests were made for a given invocation of
auracle.