Fix more clang-tidy warnings, commit to fixing them in the future
We can't do this directly in meson for a bunch of reasons:

1) meson insists on stuffing -pipe into the compile commands. that's
fine, but clang-tidy chokes on it and there doesn't seem to be a way
around it.
2) we have to provide the list of files to be linted to clang-tidy, and
meson provides no way of filtering lists of files to only pass the .cc
files (without the headers).

So, build a script which looks at git-tracked files and runs clang-tidy
for us. Wire that up through 'ninja tidy'.