Rework all of the request URL building
Overhaul our querystring building by changing the algorithm
substantially:

1) Get rid of the distinction between "params" and "args". Require that
all non "arg" or "arg[]" parameters be provided up front to the request
constructor and build the base of the URI at once.
2) Get rid of the querystring class and build a single URI within the
RpcRequest class. Chop it up as needed. Borrow some real string
functions with inspiration from absl.

The algorithm to chop up too-long querystrings looks backwards from the
end of the assembled querystring and hops through ampersands until the
querystring is short enough. This favors the typical cases where the
number of args is small enough to fit in a single querystring.

There's a fair bit of test churn in this commit because we move away
from using a std::map and to just a std::vector of string pairs. That,
however, makes it easier to reason about the ordering of params for
the future.