Extend buildorder output
This should help to address a few user requests:

1) buildorder shouldn't consider what's installed and omit these
packages. Instead, we tag output to describe what steps (if any) need to
be taken. Output is either REPOS (available for install via pacman), or
AUR (can be built from the AUR). Either of these can be prefixed with
SAITSFIED to denote that the dependency is already installed locally. As
a last resort, a depedency might be UNKNOWN when the package isn't found
anywhere (#4).

2) buildorder should include the target package(s) in the output. (#32)

3) For packages which come from the AUR, the pkgbase will be emitted as
a third column, e.g.

  AUR systemd-libs-git systemd-git

This allows for deduping and also knowing the directory the
cloned/downloaded package will be found in (#9)

I'll no doubt gain a few new bug reports with this change...