Remove IgnorePkg handling
Generally, if folks are using an AUR helper to download a specific
package, they don't care if it's being ignored by pacman or pacman
wrappers.

See #44 for further discussion.

[dreisner: there's also a lot of edge cases here, and a lot of
conditions that are just hard to get right. There doesn't seem to be a
default that pleases everyone, and adding more flags/config around this
isn't the direction I want to go in.]