Ignore pacman repos in package recursion and sync operations
In #12, users are asking for the --ignorerepo flag to come back, but
what they really want is for Auracle to make the AUR take precedence
over packages known to pacman. This was unwieldy and terribly slow in
cower, but Auracle is fast and not terrible. This commit does two
things:

* Skips discovering the notion of "foreign" packages when calling
  'sync'. We can just send the whole DB to the AUR since Auracle
  understands how to chunk large requests.
* Skips filtering out of packages known to pacman when recursing through
  dependencies. We only error out, suggesting the package is unknown,
  only after both the AUR and pacman (in that order) deny the existence
  of the package.

This should make Auracle play more nicely with unofficial DBs that
contain packages also known to the AUR. Really should add some tests for
this...

Fixes #12.