Revert addition of json-based aur::Package constructor
This walks back a change from 5e7bcad26 which seems to trigger a
compilation bug in gcc8 (but not gcc9 or clang). Let's just revert this
and build packages manually until we come up with a better solution.
It's not clear extending the real API just for the sake of tests is a
good idea anyways.