simplify handling of event loop exit
We have a finite lifetime on our event loop and can't consistently use
sd_event_exit() to terminate it. Instead, let's just use a boolean flag
to indicate cancellation and act on that in Aur::Wait. This avoids the
nonsense of sd_event_get_exit_code failing with -ENODATA in our success
case because we never called sd_event_exit.