Drop distinction between "raw" and "tarball" requests
The net effect here is that we stop sending "Accept-Encoding: identity"
headers for tarball requests, meaning that tarballs are now
double-compressed (once for the .tar.gz, again for the transfer
encoding). That should be fine, since tarballs are generally expected to
be quite small, and I'll probably drop support for fetching them in
auracle in the future.