unbreak meson wrap for fmt due to nonexisting domain
The wrapdb is a pile of... unreliability, and it keeps going down. The
only thing it does anyway is serve as a redirection service to the files
which are really stored on Github, but "we need our own domain, it's
more reliable if we ever move away from Github". The problem is that
this logic is all backwards -- the custom domain is the unreliable thing
here. In the latest instance:

- "It seems that I lost that server entirely."
- "... you lost it?!"
- "well, it was a VM on a server of my ex-ex-ex-ex-boss and apparently he
  did something with it and I don't have access there anymore."

So let's not trust wrapdb to ever be reliable, and instead download
wrap patches from github which is very reliable. In the supremely
unlikely situation that meson not only moves away from github but nukes
all their old data from high orbit, we can just update this again.