Fix infinite loop in AurImpl::Wait
Fixes #73

When receiving CURL_POLL_REMOVE from libcurl, we calls FinishRequest
with the sd_event_source handle, which would try to remove the
sd_event_source handle from active_requests_. Yet the only
sd_event_source handles stored in active_requests_ are the handles for
git child processes. Essentially we are trying to remove the wrong
handles, which leave completed requests in active_requests_, causing
Wait() to loop forever (or until timeout).

This commits add a call to CheckFinished() in that case to make sure no
completed curl requests are left in active_requests_, which fixes this
problem.