Search by longest found candidate in regex searches
Given a string like "one[two]three", we previously stopped after finding
"one", as it was a string with length of at least 2 characters. However,
querying the AUR with very short strings is non-ideal, as it increases
our likelihood of exceeding the result count.

Of course, this new approach could just as easily backfire on us, as
well. Consider the case of 'vim.*-git' which works before this change
because we search for "vim", and is broken after because we search for
"-git" (which has too many results).

Add unit tests galore so that I can motivate myself towards rewriting
this in terms of std::string_view and maybe improve it further.