Add "resolve" verb to do dependency resolution
This takes in inputs like "curl=8.0.1" or "curl<10" and returns only the
packages that can *actually* satisfy the dependency. This leverages the
"provider" searchby functionality to return candidates that can be
post-filtered to determine if they meet the requirement.

This is a small but significant step towards properly supporting
versioned dependencies as requested in #3. In order to support this in
recursive clones and 'buildorder', we're likely due for some significant
surgery on the PackageCache.