Create a wrapper for Resolve in auracle to simplify logic
This makes the frontend code waaaaaaay simpler, as resolving is already
pretty complicated. This also makes it easy to let resolve accept
multiple arguments, as non-sensical as that may end up being for the
caller.

Note that there's also a change in behavior here -- resolve now behaves
more like search WRT exit statuses. Lack of results is not an error.